requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
attrs==25.4.0
bidict==0.23.1
braceexpand==0.1.7
certifi==2025.11.12
charset-normalizer==3.4.4
circuitsvis==1.43.3
click==8.3.1
clip==0.2.0
contourpy==1.3.3
cycler==0.12.1
datasets==4.4.1
debugpy==1.8.19
dictionary-learning==0.1.0
diffusers==0.36.0
dill==0.4.0
einops==0.8.1
filelock==3.20.1
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.11
ImageHash==4.3.2
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==9.8.0
jaxtyping==0.3.4
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.7.0
kiwisolver==1.4.9
lightning-utilities==0.15.2
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mmsegmentation==1.2.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
narwhals==2.14.0
networkx==3.6.1
nnsight==0.3.7
numba==0.63.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
open_clip_torch==3.2.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.5
python-engineio==4.12.3
python-socketio==5.15.0
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentencepiece==0.2.1
sentry-sdk==2.47.0
simple-websocket==1.1.0
smmap==5.0.2
StrEnum==0.4.15
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.22
tokenizers==0.22.1
torch==2.9.1
torchmetrics==1.8.2
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
typing-inspection==0.4.2
tzdata==2025.3
umap-learn==0.5.9.post2
urllib3==2.6.2
wadler_lindig==0.1.7
wandb==0.23.1
wcwidth==0.2.14
webdataset==1.0.2
websocket-client==1.9.0
wsproto==1.3.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
